Week_4_RAG/01_Building_An_LLM_Application/01_Using_LLMs/requirements.txt
llama-index-llms-groq
python-dotenv
llama-index
groq